refactor: move getDataFromExternalSources
